define miner task and wire up scheduler
